\name{RcmdrPlugin.StatisticalURV-internal}
\title{Internal RcmdrURVPlugin objects}
\alias{oneWayANOVA}
\alias{OnPosthoc}
\alias{OnOptions}
\alias{OnGraphics}
\alias{twoWayANOVA}
\alias{OnModelsTwoWayANOVA}
\alias{OnGraphicsTwoWayANOVA}
\alias{OnCancel}
\alias{OnHelp}
\description{Internal RcmdrPlugin.StatisticalURV objects.}
\details{These are not to be called by the user.}
\keyword{internal}
