\name{RcmdrPlugin.StatisticalURV-package}
\alias{RcmdrPlugin.StatisticalURV-package}
\alias{RcmdrPlugin.StatisticalURV}
\docType{package}
\title{
Some statistical models dialog to the Rcmdr package
}
\description{
This package provides an Rcmdr "plug-in". It provides a one way ANOVA dialog, ...
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.StatisticalURV\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2012-12-06\cr
License: \tab GPL-3\cr
}
}

\author{
Daniela Vicente <daniela.vicente@estudiants.urv.cat>

Maintainer: Daniela Vicente <daniela.vicente@estudiants.urv.cat>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

